Moved client net rec into segment.

+++ src/common/gameplay/game/game_state.rs
@@ -59,7 +59,7 @@ impl GameState{
 
         self.world.update_entities(&mut self.storages, pending);
     }
-    pub fn simulate_tick(&mut self, sim_info: InfoForSim, delta: f64){
+    pub fn simulate_tick(&mut self, sim_info: InfoForSim, delta: f32){
         for (player_id, input) in &sim_info.inputs_map{
             if input.new_player{
                 self.init_new_player(*player_id);
